+++ fury/tests/test_interpolators.py
@@ -1,18 +1,10 @@
-import time
 from itertools import combinations
-
-import numpy as np
 import numpy.testing as npt
-
-from fury import actor
-import fury.animation.helpers as helpers
 from fury.animation.interpolator import linear_interpolator, \
     step_interpolator, cubic_spline_interpolator, cubic_bezier_interpolator, \
     spline_interpolator, hsv_color_interpolator, lab_color_interpolator, \
     xyz_color_interpolator, slerp
-from fury.animation.timeline import Timeline
 from fury.testing import *
-from fury.ui import PlaybackPanel
 
 
 def assert_not_equal(x, y):
@@ -216,145 +208,3 @@ def test_slerp():
     for timestamp in range(-100, 100, 1):
         timestamp /= 10
         interp_slerp(timestamp)
-
-
-def test_timeline():
-    shaders = False
-    tl = Timeline(playback_panel=Timeline)
-    tl.set_position(0, np.array([1, 1, 1]))
-    # overriding a keyframe
-    tl.set_position(0, np.array([0, 0, 0]))
-    tl.set_position(3, np.array([2, 2, 2]))
-    tl.set_position(5, np.array([3, 15, 2]))
-    tl.set_position(7, np.array([4, 2, 20]))
-
-    tl.set_opacity(0, 0)
-    tl.set_opacity(7, 1)
-
-    tl.set_rotation(0, np.array([90, 0, 0]))
-    tl.set_rotation(7, np.array([0, 180, 0]))
-
-    tl.set_scale(0, np.array([1, 1, 1]))
-    tl.set_scale(7, np.array([5, 5, 5]))
-
-    tl.set_color(0, np.array([1, 0, 1]))
-
-    # test playback panel
-    assert_true(isinstance(tl.playback_panel, PlaybackPanel))
-
-    for t in [-10, 0, 2.2, 7, 100]:
-        tl.seek(t)
-        assert_less_equal(tl.current_timestamp, tl.final_timestamp)
-        assert_greater_equal(tl.current_timestamp, 0)
-
-        assert_greater_equal(tl.current_timestamp,
-                             tl.playback_panel.current_time)
-
-        if 0 <= t <= tl.final_timestamp:
-            npt.assert_almost_equal(tl.current_timestamp, t)
-            # check if seeking a certain time affects the time slider's value.
-            npt.assert_almost_equal(tl.current_timestamp,
-                                    tl.playback_panel.current_time)
-
-    tl.play()
-    t_before = tl.current_timestamp
-    time.sleep(0.1)
-    assert_not_equal(tl.current_timestamp, t_before)
-    assert_true(tl.playing)
-
-    tl.pause()
-    t_before = tl.current_timestamp
-    assert_true(tl.paused)
-    time.sleep(0.1)
-    npt.assert_almost_equal(tl.current_timestamp, t_before)
-
-    tl.stop()
-    assert_true(tl.stopped)
-    npt.assert_almost_equal(tl.current_timestamp, 0)
-
-    npt.assert_almost_equal(tl.get_position(0), np.array([0, 0, 0]))
-    npt.assert_almost_equal(tl.get_position(7), np.array([4, 2, 20]))
-
-    tl.set_position_interpolator(linear_interpolator)
-    tl.set_position_interpolator(cubic_bezier_interpolator)
-    tl.set_position_interpolator(step_interpolator)
-    tl.set_position_interpolator(cubic_spline_interpolator)
-    tl.set_position_interpolator(spline_interpolator, degree=2)
-    tl.set_rotation_interpolator(step_interpolator)
-    tl.set_scale_interpolator(linear_interpolator)
-    tl.set_opacity_interpolator(step_interpolator)
-    tl.set_color_interpolator(linear_interpolator)
-
-    npt.assert_almost_equal(tl.get_position(0), np.array([0, 0, 0]))
-    npt.assert_almost_equal(tl.get_position(7), np.array([4, 2, 20]))
-
-    npt.assert_almost_equal(tl.get_color(7), np.array([1, 0, 1]))
-    tl.set_color(25, np.array([0.2, 0.2, 0.5]))
-    assert_not_equal(tl.get_color(7), np.array([1, 0, 1]))
-    assert_not_equal(tl.get_color(25), np.array([0.2, 0.2, 0.5]))
-
-    cube = actor.cube(np.array([[0, 0, 0]]))
-    tl.add_actor(cube)
-
-    # using force since the animation is not playing
-    tl.update_animation(force=True)
-
-    if not shaders:
-        npt.assert_almost_equal(tl.get_position(tl.current_timestamp),
-                                cube.GetPosition())
-        npt.assert_almost_equal(tl.get_scale(tl.current_timestamp),
-                                cube.GetScale())
-        npt.assert_almost_equal(tl.get_rotation(tl.current_timestamp),
-                                cube.GetOrientation())
-
-
-def test_helper_functions():
-    keyframes = {
-        0: {'value': np.array([0, 0, 0])},
-        1: {'value': np.array([1, 0, 0])},
-        2: {'value': np.array([2, 0, 0])},
-    }
-    # Test `get_timestamps_from_keyframes`
-    timestamps = helpers.get_timestamps_from_keyframes(keyframes)
-    npt.assert_equal(len(timestamps), len(keyframes))
-    npt.assert_equal(np.array(sorted(keyframes.keys())), timestamps)
-
-    values = helpers.get_values_from_keyframes(keyframes)
-    npt.assert_array_equal(values, np.array([i['value'] for i in
-                                             keyframes.values()]))
-    distance = helpers.get_distances(values)
-    expected_distances = np.array([1, 1])
-    npt.assert_equal(distance, expected_distances)
-
-    for t in range(-100, 100, 1):
-
-        next_ts = helpers.get_next_timestamp(timestamps, t / 10)
-        npt.assert_(next_ts in timestamps, "Timestamp is not valid")
-        next_ts_2 = helpers.get_next_timestamp(timestamps, t / 10,
-                                               include_first=True)
-        npt.assert_(next_ts_2 in timestamps, "Timestamp is not valid")
-
-        prev_ts = helpers.get_previous_timestamp(timestamps, t / 10)
-        npt.assert_(prev_ts in timestamps, "Timestamp is not valid")
-        prev_ts_2 = helpers.get_previous_timestamp(timestamps, t / 10,
-                                                   include_last=True)
-        npt.assert_(prev_ts_2 in timestamps, "Timestamp is not valid")
-
-        assert_greater(next_ts, prev_ts)
-        assert_greater_equal(next_ts, prev_ts_2)
-        assert_greater_equal(next_ts_2, prev_ts_2)
-        assert_greater_equal(next_ts, prev_ts_2)
-        assert_not_equal(next_ts, prev_ts)
-
-        for i in range(-100, 100, 1):
-            i /= 10
-            tt = helpers.get_time_tau(i, prev_ts, next_ts)
-            assert_greater_equal(tt, 0)
-            assert_less_equal(tt, 1)
-
-            # lerp
-            v1 = keyframes.get(prev_ts).get('value')
-            v2 = keyframes.get(next_ts).get('value')
-
-            interp_value = helpers.lerp(v1, v2, prev_ts, next_ts, i)
-            assert_arrays_equal(tt * (v2-v1) + v1, interp_value)
